+++ scripts/integration-helpers/framework.ts
@@ -1,4 +1,4 @@
-import * as request from "request";
+import fetch, { Response } from "node-fetch";
 
 import { CLIProcess } from "./cli";
 
@@ -124,35 +124,30 @@ export class TriggerEndToEndTest {
     return this.cliProcess ? this.cliProcess.stop() : Promise.resolve();
   }
 
-  invokeHttpFunction(name: string, zone = FIREBASE_PROJECT_ZONE): Promise<request.Response> {
+  invokeHttpFunction(name: string, zone = FIREBASE_PROJECT_ZONE): Promise<Response> {
     const url = `http://localhost:${[this.functionsEmulatorPort, this.project, zone, name].join(
       "/"
     )}`;
-    return new Promise((resolve, reject) => {
-      request.get(url, {}, (err, res) => {
-        if (err) return reject(err);
-        resolve(res);
-      });
-    });
+    return fetch(url);
   }
 
-  writeToRtdb(): Promise<request.Response> {
+  writeToRtdb(): Promise<Response> {
     return this.invokeHttpFunction("writeToRtdb");
   }
 
-  writeToFirestore(): Promise<request.Response> {
+  writeToFirestore(): Promise<Response> {
     return this.invokeHttpFunction("writeToFirestore");
   }
 
-  writeToPubsub(): Promise<request.Response> {
+  writeToPubsub(): Promise<Response> {
     return this.invokeHttpFunction("writeToPubsub");
   }
 
-  writeToAuth(): Promise<request.Response> {
+  writeToAuth(): Promise<Response> {
     return this.invokeHttpFunction("writeToAuth");
   }
 
-  writeToScheduledPubsub(): Promise<request.Response> {
+  writeToScheduledPubsub(): Promise<Response> {
     return this.invokeHttpFunction("writeToScheduledPubsub");
   }