+++ src/extensions/emulator/optionsHelper.ts
@@ -1,7 +1,7 @@
 import * as fs from "fs-extra";
 import * as _ from "lodash";
 import * as path from "path";
-import * as paramHelper from "./paramHelper";
+import * as paramHelper from "../paramHelper";
 import * as specHelper from "./specHelper";
 import * as triggerHelper from "./triggerHelper";
 import { Resource } from "../extensionsApi";
@@ -131,9 +131,9 @@ function buildConfig(
  * values for properties.sourceDirectory.
  * @param functionResources An array of function type resources
  */
-function getFunctionSourceDirectory(functionResources: Resource[]): String {
+function getFunctionSourceDirectory(functionResources: Resource[]): string {
   let sourceDirectory;
-  for (let r of functionResources) {
+  for (const r of functionResources) {
     let dir = _.get(r, "properties.sourceDirectory");
     if (!dir) {
       EmulatorLogger.forEmulator(Emulators.FUNCTIONS).log(