+++ src/test/utils.spec.ts
@@ -110,16 +110,31 @@ describe("utils", () => {
   });
 
   describe("getDatabaseViewDataUrl", () => {
-    it("should get a view data url for prod", () => {
+    it("should get a view data url for legacy prod URL", () => {
       expect(
-        utils.getDatabaseViewDataUrl("https://firebaseio.com", "fir-proj", "/foo/bar")
-      ).to.equal("https://console.firebase.google.com/project/fir-proj/database/data/foo/bar");
+        utils.getDatabaseViewDataUrl("https://firebaseio.com", "fir-proj", "fir-ns", "/foo/bar")
+      ).to.equal(
+        "https://console.firebase.google.com/project/fir-proj/database/fir-ns/data/foo/bar"
+      );
+    });
+
+    it("should get a view data url for new prod URL", () => {
+      expect(
+        utils.getDatabaseViewDataUrl(
+          "https://firebasedatabase.app",
+          "fir-proj",
+          "fir-ns",
+          "/foo/bar"
+        )
+      ).to.equal(
+        "https://console.firebase.google.com/project/fir-proj/database/fir-ns/data/foo/bar"
+      );
     });
 
     it("should get a view data url for the emulator", () => {
       expect(
-        utils.getDatabaseViewDataUrl("http://localhost:9000", "fir-proj", "/foo/bar")
-      ).to.equal("http://localhost:9000/foo/bar.json?ns=fir-proj");
+        utils.getDatabaseViewDataUrl("http://localhost:9000", "fir-proj", "fir-ns", "/foo/bar")
+      ).to.equal("http://localhost:9000/foo/bar.json?ns=fir-ns");
     });
   });
 
@@ -231,6 +246,16 @@ describe("utils", () => {
       expect(utils.datetimeString(new Date("February 7, 2020 8:01:01-07:00"))).to.match(
         /^2020-02-07 \d\d:01:01$/
       );
+    });
+  });
+
+  describe("streamToString/stringToStream", () => {
+    it("should be able to create and read streams", async () => {
+      const stream = utils.stringToStream("hello world");
+      if (!stream) {
+        throw new Error("stream came back undefined");
+      }
+      await expect(utils.streamToString(stream)).to.eventually.equal("hello world");
     });
   });
 });